Tools to compute the number of samples and expected values
Methods to compute the number of samples, the distance between two
vertex pairs, and the expected values of time and quality in LKDPI are
implemented. The distance function is used to compute the Lipschitz
bound (not implemented yet) which will then be used to compute inputs
find the number of samples.